+++ dbms/src/Columns/ColumnFixedString.cpp
@@ -260,14 +260,14 @@ ColumnPtr ColumnFixedString::permute(const Permutation & perm, UInt64 limit) con
 }
 
 
-ColumnPtr ColumnFixedString::index(const IColumn & indexes, UInt64 limit) const
+ColumnPtr ColumnFixedString::index(const IColumn & indexes, size_t limit) const
 {
     return selectIndexImpl(*this, indexes, limit);
 }
 
 
 template <typename Type>
-ColumnPtr ColumnFixedString::indexImpl(const PaddedPODArray<Type> & indexes, UInt64 limit) const
+ColumnPtr ColumnFixedString::indexImpl(const PaddedPODArray<Type> & indexes, size_t limit) const
 {
     if (limit == 0)
         return ColumnFixedString::create(n);
